// src/core/domain/schema.ts
export const LANGUAGES = ['en', 'pl'] as const;
export const LanguageSchema = v.enum(LANGUAGES);
export type Language = (typeof LANGUAGES)[number];